requirements.txt

python-dotenv
chainlit
litellm
